[chain]refactor create_block_template.
