+++ core/traits/src/lib.rs
@@ -7,7 +7,7 @@ mod chain_state;
 pub mod mock;
 mod txpool_service;
 
-pub use chain::{AsyncChain, Chain, ChainReader, ChainWriter};
+pub use chain::{Chain, ChainReader, ChainWriter};
 pub use chain_service::{ChainAsyncService, ChainService};
 pub use chain_state::{ChainState, ChainStateReader, ChainStateWriter};
 pub use txpool_service::TxPoolAsyncService;
